TD2: Broken out characteristics, api tools, docstring and minor fixes.
