+++ src/main.py
@@ -1,43 +1,16 @@
-import requests
-import pandas as pd
-
+"""Main point of entry."""
+import logging
 import configparser
 import json
 import os
 import datetime as dt
 
-from src.utilities.config_tools import *
+import pandas as pd
+
+import src.utilities.config_tools as ct
 import src.utilities.json_tools as jt
-
-
-def raise_request(url: str, headers: dict = None) -> requests.Response:
-    """
-
-    Parameters
-    ----------
-    url : str
-        Address to access
-    headers : dict
-        Header values to pass into "requests.get" to alter the response from
-        the url
-        The default is None
-
-    Returns
-    -------
-    requests.Response
-        Response object for url to pull data etc. from.
-
-    """
-    r = requests.get(url, headers=headers)
-
-    if r.status_code != 200:
-        logging.warning('Error code {0} from {1}. Processes will '
-                        'attempt to continue without this request, '
-                        'but no assurance can be made about future '
-                        'states'.format(r.status_code, url))
-        r = None
-
-    return r
+import src.dataproc.characteristics as char
+from src.utilities.api_tools import raise_request
 
 
 def scrape_mp_data(config: configparser.ConfigParser):
@@ -58,7 +31,8 @@ def scrape_mp_data(config: configparser.ConfigParser):
     json_path = config['PARSER']['mp_json_path']
     url = config['PARSER']['members_url']
 
-    r = raise_request(url, headers=cfg_get_dict(config, 'PARSER', 'headers'))
+    r = raise_request(url, headers=ct.cfg_get_dict(config, 'PARSER',
+                                                   'headers'))
 
     data, r = jt.jsonify_response(r)
     jt.write_to_json(data, json_path)
@@ -150,11 +124,11 @@ def clean_mp_data(config: configparser.ConfigParser,
 
     """
     # Keep only desired columns.
-    mp_cols = cfg_get_list(config, 'PARSER', 'mp_cols')
+    mp_cols = ct.cfg_get_list(config, 'PARSER', 'mp_cols')
     core_mp_data = full_member_data[mp_cols]
 
     # Create clean name column - no titles or honourifics.
-    titles = cfg_get_list(config, 'PARSER', 'titles')
+    titles = ct.cfg_get_list(config, 'PARSER', 'titles')
     core_mp_data['clearName'] = core_mp_data['DisplayAs'].str.replace(
         '|'.join(titles), '').str.strip()
 
@@ -162,134 +136,6 @@ def clean_mp_data(config: configparser.ConfigParser,
         core_mp_data.shape, core_mp_data.columns))
 
     return core_mp_data
-
-
-def get_characteristics():
-    """
-    Reads the manually created characteristic data.
-
-    Returns
-    -------
-    pd.DataFrame
-        Characteristic data read from file
-    """
-    logging.debug('Reading characteristic data from {0}'.format(
-        config['PARSER']['characteristic_path']))
-
-    characteristic_data = pd.read_csv(config['PARSER']['characteristic_path'])
-
-    return characteristic_data
-
-
-def report_missing_characteristics(config: configparser.ConfigParser,
-                                   missing_characteristics: pd.DataFrame):
-    """
-    Takes a dataframe which is missing matching characteristics, and writes
-    the details to a file.
-
-    Writes to a missing_identifiers log csv, and optionally if config value is
-    set, can also append the known data about the unknown MP to the default
-    identifiers csv.
-
-    Parameters
-    ----------
-    config : configparser.ConfigParser
-        Contents of config file
-    missing_characteristics : pd.DataFrame
-        DataFrame of records missing from characteristic data.
-    """
-    logging.warning("Missing characteristics for {0} MPs in {1}".format(
-        len(missing_characteristics), config['PARSER']['characteristic_path']))
-
-    now = dt.datetime.now().isoformat()
-    warn_path = "logs/missing_identifiers.csv".format(now)
-    logging.warning("Writing details of missing data to {0}".format(warn_path))
-
-    key_details = missing_characteristics[['@Member_Id', 'DisplayAs', 'ListAs',
-                                           'MemberFrom', 'Party.#text']]
-    key_details.insert(0, 'timestamp', now)
-    key_details.insert(1, 'missing', 'identifiers.csv')
-
-    key_details.to_csv(warn_path, mode='a', index=False)
-
-    if config['PARSER']['append_missing_characteristics'] == "True":
-        append_details = missing_characteristics[["@Member_Id", "DisplayAs",
-                                                  "ListAs", "clearName"]]
-        append_path = config['PARSER']['characteristic_path']
-
-        logging.warning("Appending missing characteristic records to default "
-                        "characteristic data - will need expanding with "
-                        "user-filled values.")
-        append_details.to_csv(append_path, mode='a', index=False, header=False)
-
-
-def add_characteristics(config: configparser.ConfigParser,
-                        member_data: pd.DataFrame,
-                        characteristic_data: pd.DataFrame):
-    """
-    Appends characteristics to members, where possible.
-
-    Will first merge on columns specified in characteristic_merge in config,
-    where failures occur will attempt to further merge on columns in
-    characteristic_merge_excess.
-
-    Resulting data will include successful matches from the first join,
-    and all records from the result of the second - successes OR failures -
-    note that this might mean some rows are missing characteristics. They
-    will be logged and handled in report_missing_characteristics.
-
-    Parameters
-    ----------
-    config : configparser.ConfigParser
-        Contents of config file
-    member_data : pd.DataFrame
-        MP member data from either scraped gov website, or local copy.
-    characteristic_data : pd.DataFrame
-        Characteristic data from compiled identifiers.csv file
-
-    Returns
-    -------
-    pd.DataFrame:
-        MP data with matching characteristic data where possible. If no
-        matching characteristic data will be found, MP will be included, but
-        characteristic columns will be blank
-    """
-
-    logging.debug('Adding characteristic data to core data')
-
-    default_columns = cfg_get_list(config, 'PARSER', 'characteristic_merge')
-    backup_columns = cfg_get_list(config, 'PARSER',
-                                  'characteristic_merge_excess')
-
-    primary_merge = pd.merge(member_data, characteristic_data,
-                             on=default_columns, how='left', indicator=True)
-
-    # Split matched + unmatched out.
-    missing_characteristics = primary_merge[primary_merge['_merge'] ==
-                                            'left_only']
-    primary_merge = primary_merge[primary_merge['_merge'] == 'both']
-
-    # Get original data where failed to match
-    missing_member_data = member_data[member_data['@Member_Id'].isin(
-        missing_characteristics['@Member_Id'])]
-
-    # Drop columns which will duplicate for backup merge
-    characteristic_data.drop(columns=['DisplayAs', 'ListAs'], inplace=True)
-
-    # Merge on backup columns
-    secondary_merge = pd.merge(missing_member_data, characteristic_data,
-                               on=backup_columns, how='left', indicator=True)
-
-    # New missing_characteristics can overwrite previous
-    missing_characteristics = (secondary_merge[
-                                secondary_merge['_merge'] == 'left_only'])
-
-    matched_members = pd.concat([primary_merge, secondary_merge])
-
-    if not missing_characteristics.empty:
-        report_missing_characteristics(config, missing_characteristics)
-
-    return matched_members
 
 
 def main(config: configparser.ConfigParser):
@@ -313,9 +159,9 @@ def main(config: configparser.ConfigParser):
 
     core_mp_data = clean_mp_data(config, full_member_data)
 
-    characteristic_data = get_characteristics()
-    core_mp_data = add_characteristics(config, core_mp_data,
-                                       characteristic_data)
+    characteristic_data = char.get_characteristics(config)
+    core_mp_data = char.add_characteristics(config, core_mp_data,
+                                            characteristic_data)
 
     core_mp_data.to_csv(config['PARSER']['full_mp_path'], header=True,
                         index=False)
@@ -329,7 +175,7 @@ def check_for_errors():
     Lists any files that exist, the last modification time, and the number of
     records contained within.
     """
-    files = cfg_get_list(config, 'ERROR_CHECKING', 'files')
+    files = ct.cfg_get_list(config, 'ERROR_CHECKING', 'files')
 
     files = [file for file in files if os.path.exists(file)]
