+++ src/utilities/config_tools.py
@@ -1,4 +1,4 @@
-# Contains methods for dealing with config files.
+"""Contains methods for dealing with config files."""
 import configparser
 import logging
 from typing import Callable
